# openapi.yaml
openapi: 3.0.3
info:
  title: PartyHub API
  description: Social event platform for organizing and discovering parties, meetups, and social gatherings
  version: 1.0.0
  contact:
    name: PartyHub Team
servers:
  - url: http://localhost:8080
    description: Development server
paths:
  /api/parties:
    get:
      summary: List, filter, or sort parties
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Search query (searches title and description)
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category name
        - name: date_from
          in: query
          schema:
            type: string
          description: Start date (dd-mm-YYYY or dd.MM.yyyy HH:mm)
        - name: date_to
          in: query
          schema:
            type: string
          description: End date (dd-mm-YYYY or dd.MM.yyyy HH:mm)
        - name: sort
          in: query
          schema:
            type: string
          description: Sort field (asc or desc)
        - name: X-User-Id
          in: header
          schema:
            type: integer
            format: int64
          description: User ID for authentication
      responses:
        '200':
          description: List of parties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Party'
    post:
      summary: Create a new party
      parameters:
        - name: user
          in: query
          schema:
            type: integer
            format: int64
        - name: X-User-Id
          in: header
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartyCreateDto'
      responses:
        '200':
          description: Party created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Party'
  /api/parties/{id}:
    get:
      summary: Get party by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: user
          in: query
          schema:
            type: integer
            format: int64
        - name: X-User-Id
          in: header
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Party details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Party'
        '404':
          description: Party not found
    put:
      summary: Update party
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: user
          in: query
          schema:
            type: integer
            format: int64
        - name: X-User-Id
          in: header
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartyCreateDto'
      responses:
        '200':
          description: Party updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Party'
    delete:
      summary: Delete party
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Party deleted
  /api/parties/{id}/join:
    post:
      summary: Join party
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: X-User-Id
          in: header
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Joined party
    delete:
      summary: Leave party
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: X-User-Id
          in: header
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Left party
  /api/parties/{id}/join/status:
    get:
      summary: Check join status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: X-User-Id
          in: header
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Join status
  /api/parties/{id}/can-edit:
    get:
      summary: Check if user can edit party
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: user
          in: query
          schema:
            type: integer
            format: int64
        - name: X-User-Id
          in: header
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Can edit status
  /api/parties/{id}/locations:
    get:
      summary: Get party attendee locations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of attendee locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserLocation'
  /api/parties/{id}/media:
    get:
      summary: Get party media
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of party media
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Media'
    post:
      summary: Upload party media
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: user
          in: query
          schema:
            type: integer
            format: int64
        - name: X-User-Id
          in: header
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Media uploaded
...
  /api/invitations:
    post:
      summary: Create invitation
      parameters:
        - name: user
          in: query
          schema:
            type: integer
            format: int64
        - name: X-User-Id
          in: header
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationDto'
      responses:
        '200':
          description: Invitation created
    get:
      summary: Get invitations
      parameters:
        - name: direction
          in: query
          schema:
            type: string
            enum: [received, sent]
          description: Filter by direction (received or sent)
        - name: user
          in: query
          schema:
            type: integer
            format: int64
        - name: X-User-Id
          in: header
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of invitations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Invitation'
  /api/invitations/{id}:
    delete:
      summary: Delete invitation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: user
          in: query
          schema:
            type: integer
            format: int64
        - name: X-User-Id
          in: header
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Invitation deleted
...
components:
  schemas:
    Party:
      type: object
      properties:
        id:
          type: integer
          format: int64
        host_user:
          $ref: '#/components/schemas/User'
        category:
          $ref: '#/components/schemas/Category'
        title:
          type: string
        time_start:
          type: string
          format: date-time
        time_end:
          type: string
          format: date-time
        max_people:
          type: integer
        min_age:
          type: integer
        max_age:
          type: integer
        website:
          type: string
        description:
          type: string
        fee:
          type: number
          format: double
        created_at:
          type: string
          format: date-time
        location:
          $ref: '#/components/schemas/Location'
        visibility:
          type: string
          enum: [PUBLIC, PRIVATE]
    PartyCreateDto:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        fee:
          type: number
          format: double
        time_start:
          type: string
          format: date-time
        time_end:
          type: string
          format: date-time
        max_people:
          type: integer
        min_age:
          type: integer
        max_age:
          type: integer
        website:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        location_address:
          type: string
        category_id:
          type: integer
          format: int64
        theme:
          type: string
        visibility:
          type: string
          enum: [PUBLIC, PRIVATE]
        selectedUsers:
          type: array
          items:
            type: integer
            format: int64
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        displayName:
          type: string
        distinctName:
          type: string
        email:
          type: string
        biography:
          type: string
...
    Location:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        address:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
    Category:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        color:
          type: string
    Media:
      type: object
      properties:
        id:
          type: integer
          format: int64
        file:
          type: string
        party:
          $ref: '#/components/schemas/Party'
    Invitation:
      type: object
      properties:
        id:
          type: integer
          format: int64
        sender:
          $ref: '#/components/schemas/User'
        recipient:
          $ref: '#/components/schemas/User'
        party:
          $ref: '#/components/schemas/Party'
    InvitationDto:
      type: object
      properties:
        recipient:
          type: integer
          format: int64
        partyId:
          type: integer
          format: int64
...
    UserLocation:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user:
          $ref: '#/components/schemas/User'
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
...